v3.0.0

+++ src/controllers/monster.controller.ts
@@ -8,6 +8,7 @@ import { readFileSync } from 'fs';
 export const get = async (req: Request, res: Response): Promise<Response> => {
   const id: Id = req.params.id;
   const monster = await Monster.query().findById(id);
+  if (!monster) return res.status(StatusCodes.NOT_FOUND).json();
   return res.status(StatusCodes.OK).json(monster);
 };
 
@@ -24,6 +25,8 @@ export const update = async (
   res: Response
 ): Promise<Response> => {
   const id: Id = req.params.id;
+  const monster = await Monster.query().findById(id);
+  if (!monster) return res.status(StatusCodes.NOT_FOUND).json();
   await Monster.query().findById(id).patch(req.body);
   return res.sendStatus(StatusCodes.OK);
 };
@@ -33,6 +36,8 @@ export const remove = async (
   res: Response
 ): Promise<Response> => {
   const id: Id = req.params.id;
+  const monster = await Monster.query().findById(id);
+  if (!monster) return res.status(StatusCodes.NOT_FOUND).json();
   await Monster.query().deleteById(id);
   return res.sendStatus(StatusCodes.NO_CONTENT);
 };
